move datetime dep to test
