+++ ksoup-test/module.yaml
@@ -16,6 +16,7 @@ test-dependencies:
     - $libs.korlibs.io
     - $libs.codepoints
     - $libs.kotlinx.coroutines.test
+    - $libs.kotlinx.datetime
 
 settings:
     kotlin:
